requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
psycopg2==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
